synapse-app/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
pillow==10.2.0
supabase>=2.24.0
qdrant-client==1.7.0
python-dotenv==1.0.0
pydantic>=2.12.4
httpx>=0.28.1
websockets>=15.0
playwright==1.40.0
beautifulsoup4==4.12.2
lxml==4.9.3
